Add `sites restore` command
The `restore` sub-command for `sites` will simply replace current test
site with a given one. It will be useful when dealing with complex bugs
or features or working on migration steps.

Let's say we need to test the migration from `stable-3.8` to
`stable-3.9`. First we would checkout `stable-3.8` in our workspace:
```
$ grdt checkout stable-3.8
```
we can add more test data to the site. Then crate a baseline:
```
$ grdt sites snapshot stable-3.8-baseline
```
Now we can switch back to `stable-3.9`:
```
$ grdt checkout stable-3.9
```
This will also change the test site to 3.9, so we need to restore 3.8
version:
```
$ grdt sites switch stable-3.8
```
Now we can do the migration testing. When we'd need to start from
scratch we can just do:
```
$ grdt sites restore stable-3.8-baseline
```
which will bring back all of the initial test data and configurations.

Change-Id: Ide5e31cb8858a69a7dc904285fa0451f95e886f5