Fix version negotiation
When negotiating git version for given plugin, we need to get the latest
version present in currently checked out branch. For this purpose we
used `git branch --contains $version` command. But when it's used
without a branch name, it will list all local branches that contains
that version. We only want to use it as a true/false value, for that we
need to always add `HEAD` as last argument.

Now, this command will either return current branch name (which should
be a _detached HEAD_) or and empty output. Making the version
negotiation work correctly again.

Change-Id: Ib1b95e3eda40a1406df743829282956a3c33a3fb